Separated out AppExceptions into multiple files for readability
